feat(offchain): split reader and sender, use different design pattern for `react`
