+++ offchain/dave-compute/src/main.rs
@@ -1,4 +1,4 @@
-use cartesi_compute_core::arena::{Arena, ArenaConfig};
+use cartesi_compute_core::arena::{ArenaConfig, EthArenaSender, StateReader};
 use cartesi_compute_core::machine::{CachingMachineCommitmentBuilder, MachineFactory};
 use cartesi_compute_core::strategy::{gc::GarbageCollector, player::Player};
 use ethers::types::Address;
@@ -22,7 +22,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         web3_private_key: private_key,
     };
 
-    let arena = Arena::new(arena_config)?;
+    let reader = StateReader::new(arena_config.clone())?;
+    let sender = EthArenaSender::new(arena_config)?;
 
     let machine_path = std::env::var("MACHINE_PATH").expect("MACHINE_PATH is not set");
     // String::from("/root/permissionless-arbitration/lua_node/program/simple-program");
@@ -35,26 +36,25 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let root_tournament = Address::from_str("0xcafac3dd18ac6c6e92c921884f9e4176737c052c")?;
 
     let mut player = Player::new(
-        arena.clone(),
         machine_factory.clone(),
         machine_path.clone(),
         CachingMachineCommitmentBuilder::new(machine_factory, machine_path),
         root_tournament.clone(),
     );
 
-    let mut gc = GarbageCollector::new(arena.clone(), root_tournament.clone());
+    let mut gc = GarbageCollector::new(root_tournament.clone());
 
     loop {
-        let tournament_states = arena.fetch_from_root(root_tournament).await?;
-        let res = player.react(tournament_states).await?;
+        let tournament_states = reader.fetch_from_root(root_tournament).await?;
+        let res = player.react(&sender, tournament_states).await?;
         if let Some(r) = res {
             info!("Tournament finished, {:?}", r);
             break;
         }
         tokio::time::sleep(Duration::from_secs(1)).await;
 
-        let tournament_states = arena.fetch_from_root(root_tournament).await?;
-        gc.react(tournament_states).await?;
+        let tournament_states = reader.fetch_from_root(root_tournament).await?;
+        gc.react(&sender, tournament_states).await?;
         tokio::time::sleep(Duration::from_secs(1)).await;
     }
